requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
mysqlclient==2.2.7
python-dotenv==1.1.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
graphene-django==3.2.3
django-graphql-jwt==0.4.0 